+++ src/BuddyClimb/Plugin.cs
@@ -1,5 +1,7 @@
 using BepInEx;
 using BepInEx.Logging;
+using BuddyClimb.Configuration;
+using BuddyClimb.Debugging;
 using HarmonyLib;
 
 namespace BuddyClimb;
@@ -15,9 +17,22 @@ public partial class Plugin : BaseUnityPlugin
     private void Awake()
     {
         Log = Logger;
+        BuddyClimbConfig.Bind(Config);
+        BuddyClimbConfig.EnableHotReload(Config);
 
         new Harmony(Id).PatchAll(typeof(Plugin).Assembly);
 
         Log.LogInfo($"Plugin {Name} is loaded!");
+    }
+
+    private void Update()
+    {
+        BuddyClimbConfig.ReloadIfChanged();
+        DebugPlayerSpawner.Update();
+    }
+
+    private void OnDestroy()
+    {
+        BuddyClimbConfig.DisableHotReload();
     }
 }